Move SAME_XR to seperate util module
